Dev: Code organization, UI tweaks
